pyproject.toml
[tool.poetry]
name = "split-wiser"
version = "0.1.0"
description = "Manage your shared expenses with this application."
authors = [
    "Amogh Madan <amoghmadaan@gmail.com>",
    "Ashutosh Raj Baranwal <ashutoshrajb@gmail.com>",
    "Jai Narayan Jain <jainaryanjain2604@gmail.com>",
]
license = "MIT"
readme = "README.md"
#packages = [{include = "split_wiser"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Flask-Migrate = "^3.1.0"
flask-marshmallow = "^0.14.0"
flask-sqlalchemy = "^3.0.2"
gunicorn = "^20.1.0"
marshmallow-sqlalchemy = "^0.28.1"


[tool.poetry.group.dev.dependencies]
tox = "^3.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"

[tool.black]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
extend-exclude = '''
^.*\b(migrations)\b.*$
'''


[tool.bandit]
exclude_dirs = [
    "./.eggs",
    "./.git",
    "./.tox",
    "./build",
    "./coverage",
    "./dist",
    "./venv",
    "./__pycache__",
    "*tests*",
]
